refactor(cloudapi): Message dan Media become sub class
