+++ src/cloudApi/Message.ts
@@ -1,0 +1,28 @@
+import { MessageType, SendMessageBody } from "../@types";
+import Wa from "../Wa";
+import WaApi from "../WaApi";
+
+export default class Message extends WaApi {
+  sendMessage<T extends MessageType>({
+    type,
+    to,
+    recipient_type,
+    ...data
+  }: SendMessageBody<T>) {
+    return this.fetcher.post({
+      url: this.url.SEND_MESSAGE,
+      body: {
+        messaging_product: "whatsapp",
+        [type]: data,
+        to,
+        recipient_type,
+        type,
+      },
+    });
+  }
+  get url(){
+    return {
+      SEND_MESSAGE: `https://graph.facebook.com/${Wa.apiVersion}/${this.phoneNumberId}/messages`,
+    }
+  }
+}
