+++ test/cloudapi/media.test.ts
@@ -3,14 +3,14 @@ import { Wa } from "../../src";
 import path from "path";
 
 describe("Cloud Api Media test", () => {
-  it("can upload file", async () => {
+  it.skip("can upload file", async () => {
     const wa = new Wa({
       token: process.env.TOKEN || "",
       defaultPhoneNumberId: process.env.DEFAULT_PHONE_NUMBER_ID || "",
       accountId: process.env.ACCOUNT_ID || "",
       appId: process.env.APP_ID || "",
     });
-    const json = (await wa.cloudApi.uploadMedia(path.join(process.cwd(), "cat.jpg"))) as any;
+    const json = (await wa.cloudApi.Media.uploadMedia(path.join(process.cwd(), "cat.jpg"))) as any;
     expect(json).toHaveProperty("id")
   });
 });
